# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/brokermr810/quantdinger
  license:
    name: See repository LICENSE
  x-api-app-version: 3.0.22
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/market/types:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Return supported market types for the add-watchlist modal.
      description: 'Visibility is controlled by the centralised :mod:`app.utils.market_visibility`

        rules — primarily the ``ENABLED_MARKETS`` whitelist, with

        ``SHOW_CN_STOCK`` / ``SHOW_HK_STOCK`` kept for back-compat. The radar

        (``/api/global-market/opportunities``) and the Agent API

        (``/api/agent/v1/markets``) read from the same helper so all three

        user-facing market lists stay in lock-step.'
      tags:
      - Market
      operationId: getApiMarketTypes
      x-visibility: public
...
  /api/global-market/overview:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get global market overview including indices, forex, crypto, and commodities.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketOverview
      x-visibility: public
  /api/global-market/heatmap:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get market heatmap data for crypto, stock sectors, forex, and indices.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketHeatmap
      x-visibility: public
  /api/global-market/news:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: 'Get financial news from various sources.  Query params: lang (''cn''|''en''|''all'').'
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketNews
      x-visibility: public
  /api/global-market/calendar:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get economic calendar events with impact indicators.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketCalendar
      x-visibility: public
  /api/global-market/sentiment:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get comprehensive market sentiment indicators.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketSentiment
      x-visibility: public
  /api/global-market/adanos-sentiment:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get optional Adanos Market Sentiment for selected US stock tickers.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketAdanosSentiment
      x-visibility: public
  /api/global-market/opportunities:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Scan for trading opportunities across Crypto, US/CN/HK Stocks, and
        Forex.
      tags:
      - GlobalMarket
      operationId: getApiGlobalMarketOpportunities
      x-visibility: public
  /api/global-market/refresh:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Force refresh all market data (clears cache).
      tags:
      - GlobalMarket
      operationId: postApiGlobalMarketRefresh
      x-visibility: public
...
openapi: 3.0.3